requirements.txt
Django==2.0.6
websocket-client==0.48.0
djangorestframework==3.8.2
dj-database-url==0.5.0
psycopg2==2.7.5
